pyproject.toml
[project]
name = "sktch2facefusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.13.0",
    "annotated-doc==0.0.4",
    "anyio==4.13.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.7",
    "click==8.3.2",
    "cuda-bindings==13.2.0",
    "cuda-pathfinder==1.5.1",
    "cuda-toolkit==13.0.2",
    "diffusers==0.37.1",
    "dlib==20.0.1",
    "exceptiongroup==1.3.1",
    "face-recognition==1.3.0",
    "face-recognition-models",
    "filelock==3.25.2",
    "fsspec==2026.3.0",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.9.0",
    "idna==3.11",
    "importlib-metadata==9.0.0",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "nvidia-cublas==13.1.0.3",
    "nvidia-cuda-cupti==13.0.85",
    "nvidia-cuda-nvrtc==13.0.88",
    "nvidia-cuda-runtime==13.0.96",
    "nvidia-cudnn-cu13==9.19.0.56",
    "nvidia-cufft==12.0.0.61",
    "nvidia-cufile==1.15.1.6",
    "nvidia-curand==10.4.0.35",
    "nvidia-cusolver==12.0.4.66",
    "nvidia-cusparse==12.6.3.3",
    "nvidia-cusparselt-cu13==0.8.0",
    "nvidia-nccl-cu13==2.28.9",
    "nvidia-nvjitlink==13.0.88",
    "nvidia-nvshmem-cu13==3.4.5",
    "nvidia-nvtx==13.0.85",
    "packaging==26.0",
    "pillow==12.2.0",
    "psutil==7.2.2",
    "pygments==2.20.0",
    "pyyaml==6.0.3",
    "regex==2026.4.4",
    "requests==2.33.1",
    "rich==14.3.3",
    "safetensors==0.7.0",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "sympy==1.14.0",
    "tokenizers==0.22.2",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "tqdm==4.67.3",
    "transformers==5.5.0",
    "triton==3.6.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "urllib3==2.6.3",
    "zipp==3.23.0",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models", rev = "e67de717267507d1e9246de95692eb8be736ab61" }
